+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add experimental Admin layout (PR #371)
 * Add the [GOV.UK Frontend](https://design-system.service.gov.uk/) library to the gem (PR #398)
 * Allow linking to the Design System on component pages (PR #401)
+* Add govuk:analytics:organisations meta tag if the current page is an organisation (PR #397)
 
 ## 9.3.6